add rtk query, async reject andpoints, refactor all pages
