+++ extractedTranslations/en/profile.json
@@ -16,6 +16,7 @@
   "Попробуйте обновить страницу": "",
   "Произошла непредвиденная ошибка": "",
   "Произошла ошибка при загрузке профиля": "",
+  "Профиль не найден!": "",
   "Редактировать": "",
   "Сохранить": "",
   "Страница профиля": "",
