+++ extractedTranslations/ru/translation.json
@@ -60,8 +60,10 @@
   "Пример": "Пример",
   "Произошла непредвиденная ошибка": "Произошла непредвиденная ошибка",
   "Произошла ошибка при загрузке профиля": "Произошла ошибка при загрузке профиля",
+  "Произошла ошибка при загрузке статей": "Произошла ошибка при загрузке статей",
   "Произошла ошибка при загрузке статьи": "Произошла ошибка при загрузке статьи",
   "Профиль": "Профиль",
+  "Профиль не найден!": "Профиль не найден!",
   "Редактировать": "Редактировать",
   "Рекомендуем": "Рекомендуем",
   "Создать статью": "Создать статью",
