+++ src/app/providers/StoreProvider/config/store.ts
@@ -11,6 +11,7 @@ import { userReducer } from 'entities/User';
 import { uiReducer } from 'features/UI';
 import { NavigateOptions, To } from 'react-router-dom';
 import { $api } from 'shared/api/api';
+import { rtkApi } from 'shared/api/rtkApi';
 import { createReducerManager } from './reducerManager';
 import { StateSchema, ThunkExtraArg } from './StateSchema';
 
@@ -23,7 +24,8 @@ export function createReduxStore(
         counter: counterReducer,
         user: userReducer,
         sidebar: sidebarReducer,
-        ui: uiReducer
+        ui: uiReducer,
+        [rtkApi.reducerPath]: rtkApi.reducer
     };
 
     const reducerManager = createReducerManager(rootReducer);
@@ -41,7 +43,7 @@ export function createReduxStore(
                 thunk: {
                     extraArgument: extraErg
                 }
-            })
+            }).concat(rtkApi.middleware)
     });
 
     // @ts-ignore
