+++ src/entities/Profile/model/types/profile.ts
@@ -1,18 +1,6 @@
 import { Country } from 'entities/Country';
 import { Currency } from 'entities/Currency';
 
-export enum ValidateProfileError {
-    INCORRECT_FIRSTNAME = 'INCORRECT_FIRSTNAME',
-    INCORRECT_LASTNAME = 'INCORRECT_LASTNAME',
-    INCORRECT_AGE = 'INCORRECT_AGE',
-    INCORRECT_USERNAME = 'INCORRECT_USERNAME',
-    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
-    INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
-    INCORRECT_CITY = 'INCORRECT_CITY',
-    NO_DATA = 'NO_DATA',
-    SERVER_ERROR = 'SERVER_ERROR'
-}
-
 export interface Profile {
     id?: string;
     firstname?: string;
@@ -23,13 +11,4 @@ export interface Profile {
     city?: string;
     username?: string;
     avatar?: string;
-}
-
-export interface ProfileSchema {
-    data?: Profile;
-    form?: Profile;
-    isLoading: boolean;
-    error?: string;
-    readonly: boolean;
-    validateError?: ValidateProfileError[];
 }
