+++ src/features/articleRecomendationsList/index.ts
@@ -1,0 +1,1 @@
+export { ArticleRecomendationsList } from './ui/ArticleRecomendationsList/ArticleRecomendationsList';
