+++ src/features/articleRecomendationsList/ui/ArticleRecomendationsList/ArticleRecomendationsList.stories.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+
+import { ArticleRecomendationsList } from './ArticleRecomendationsList';
+
+export default {
+    title: 'features/ArticleRecomendationsList/ArticleRecomendationsList',
+    component: ArticleRecomendationsList,
+    argTypes: {
+        backgroundColor: { control: 'color' }
+    }
+} as ComponentMeta<typeof ArticleRecomendationsList>;
+
+const Template: ComponentStory<typeof ArticleRecomendationsList> = (args) => (
+    <ArticleRecomendationsList {...args} />
+);
+
+export const Normal = Template.bind({});
+Normal.args = {};
