+++ src/features/editableProfileCard/model/selectors/getProfileData/getProfileData.ts
