+++ src/features/editableProfileCard/model/selectors/getProfileError/getProfileError.test.ts
