+++ src/features/editableProfileCard/model/slice/profileSlice.test.ts
@@ -2,7 +2,10 @@ import { Country } from 'entities/Country';
 import { Currency } from 'entities/Currency';
 import { fetchProfileData } from '../services/fetchProfileData/fetchProfileData';
 import { updateProfileData } from '../services/updateProfileData/updateProfileData';
-import { ProfileSchema, ValidateProfileError } from '../types/profile';
+import {
+    ProfileSchema,
+    ValidateProfileError
+} from '../types/editableProfileCardSchema';
 import { profileActions, profileReducer } from './profileSlice';
 
 const data = {
