+++ src/features/editableProfileCard/ui/EditableProfileCard/EditableProfileCard.tsx
@@ -1,0 +1,161 @@
+import { classNames } from 'shared/lib/helpers/classNames';
+import { useTranslation } from 'react-i18next';
+import { memo, useCallback } from 'react';
+import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
+import { useSelector } from 'react-redux';
+import { useInitialEffect } from 'shared/lib/hooks/useInitialEffect/useInitialEffect';
+import { Currency } from 'entities/Currency';
+import { Country } from 'entities/Country';
+import { TextTheme, Text } from 'shared/ui/Text/Text';
+import { ProfileCard } from 'entities/Profile';
+import {
+    DynamicModuleLoader,
+    ReducerList
+} from 'shared/lib/components/DynamicModuleLoader/DynamicModuleLoader';
+import { VStack } from 'shared/ui/Stack';
+import { getProfileForm } from '../../model/selectors/getProfileForm/getProfileForm';
+import { getProfileError } from '../../model/selectors/getProfileError/getProfileError';
+import { getProfileIsLoading } from '../../model/selectors/getProfileIsLoading/getProfileIsLoading';
+import { getProfileReadonly } from '../../model/selectors/getProfileReadOnly/getProfileReadonly';
+import { getProfileValidateErrors } from '../../model/selectors/getProfileValidateErrors/getProfileValidateErrors';
+import { fetchProfileData } from '../../model/services/fetchProfileData/fetchProfileData';
+import { ValidateProfileError } from '../../model/types/editableProfileCardSchema';
+import { profileActions, profileReducer } from '../../model/slice/profileSlice';
+import { EditableProfileCardHeader } from '../EditableProfileCardHeader/EditableProfileCardHeader';
+
+interface EditableProfileCardProps {
+    className?: string;
+    id: string;
+}
+
+const reducers: ReducerList = {
+    profile: profileReducer
+};
+
+export const EditableProfileCard = memo((props: EditableProfileCardProps) => {
+    const { className, id } = props;
+    const dispatch = useAppDispatch();
+    const { t } = useTranslation('profile');
+
+    const formData = useSelector(getProfileForm);
+    const error = useSelector(getProfileError);
+    const isLoading = useSelector(getProfileIsLoading);
+    const readonly = useSelector(getProfileReadonly);
+    const validateErrors = useSelector(getProfileValidateErrors);
+
+    const validateErrorTranslations = {
+        [ValidateProfileError.INCORRECT_AGE]: t('Введите корректное значение'),
+        [ValidateProfileError.INCORRECT_CITY]: t('Поле является обязательным'),
+        [ValidateProfileError.INCORRECT_COUNTRY]: t(
+            'Поле является обязательным'
+        ),
+        [ValidateProfileError.INCORRECT_CURRENCY]: t(
+            'Поле является обязательным'
+        ),
+        [ValidateProfileError.INCORRECT_FIRSTNAME]: t(
+            'Поле является обязательным'
+        ),
+        [ValidateProfileError.INCORRECT_LASTNAME]: t(
+            'Поле является обязательным'
+        ),
+        [ValidateProfileError.INCORRECT_USERNAME]: t(
+            'Поле является обязательным'
+        ),
+        [ValidateProfileError.NO_DATA]: t('Данные не указаны'),
+        [ValidateProfileError.SERVER_ERROR]: t(
+            'Произошла непредвиденная ошибка'
+        )
+    };
+
+    useInitialEffect(() => {
+        if (id) {
+            dispatch(fetchProfileData(id));
+        }
+    });
+
+    const onChangeFirstname = useCallback(
+        (value?: string) => {
+            dispatch(profileActions.updateProfile({ firstname: value || '' }));
+        },
+        [dispatch]
+    );
+
+    const onChangeLastname = useCallback(
+        (value?: string) => {
+            dispatch(profileActions.updateProfile({ lastname: value || '' }));
+        },
+        [dispatch]
+    );
+
+    const onChangeCity = useCallback(
+        (value?: string) => {
+            dispatch(profileActions.updateProfile({ city: value || '' }));
+        },
+        [dispatch]
+    );
+
+    const onChangeAge = useCallback(
+        (value?: string) => {
+            dispatch(profileActions.updateProfile({ age: Number(value || 0) }));
+        },
+        [dispatch]
+    );
+
+    const onChangeAvatar = useCallback(
+        (value?: string) => {
+            dispatch(profileActions.updateProfile({ avatar: value || '' }));
+        },
+        [dispatch]
+    );
+
+    const onChangeUsername = useCallback(
+        (value?: string) => {
+            dispatch(profileActions.updateProfile({ username: value || '' }));
+        },
+        [dispatch]
+    );
+
+    const onChangeCurrency = useCallback(
+        (currency?: Currency) => {
+            dispatch(profileActions.updateProfile({ currency }));
+        },
+        [dispatch]
+    );
+
+    const onChangeCountry = useCallback(
+        (country?: Country) => {
+            dispatch(profileActions.updateProfile({ country }));
+        },
+        [dispatch]
+    );
+
+    return (
+        <DynamicModuleLoader reducers={reducers} removeAfterAnmount>
+            <VStack gap="8" max className={classNames('', {}, [className])}>
+                <EditableProfileCardHeader />
+                {validateErrors?.length &&
+                    validateErrors.map((err) => (
+                        <Text
+                            theme={TextTheme.ERROR}
+                            title={validateErrorTranslations[err]}
+                        />
+                    ))}
+                <ProfileCard
+                    data={formData}
+                    isLoading={isLoading}
+                    error={error}
+                    readonly={readonly}
+                    validateErrors={validateErrors}
+                    onChangeFirstname={onChangeFirstname}
+                    onChangeLastname={onChangeLastname}
+                    onChangeAge={onChangeAge}
+                    onChangeCity={onChangeCity}
+                    onChangeAvatar={onChangeAvatar}
+                    onChangeUsername={onChangeUsername}
+                    onChangeCurrency={onChangeCurrency}
+                    onChangeCountry={onChangeCountry}
+                />
+            </VStack>
+        </DynamicModuleLoader>
+    );
+});
