+++ src/features/editableProfileCard/ui/EditableProfileCardHeader/EditableProfileCardHeader.tsx
@@ -1,0 +1,79 @@
+import { useTranslation } from 'react-i18next';
+import { memo, useCallback } from 'react';
+import { classNames } from 'shared/lib/helpers/classNames';
+import { useSelector } from 'react-redux';
+import { getUserAuthData } from 'entities/User';
+import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
+import { HStack } from 'shared/ui/Stack';
+import { Button, ButtonTheme } from 'shared/ui/Button/Button';
+import { Text } from 'shared/ui/Text/Text';
+import { getProfileReadonly } from '../../model/selectors/getProfileReadOnly/getProfileReadonly';
+import { getProfileData } from '../../model/selectors/getProfileData/getProfileData';
+import { profileActions } from '../../model/slice/profileSlice';
+import { updateProfileData } from '../../model/services/updateProfileData/updateProfileData';
+
+interface EditableProfileCardHeaderProps {
+    className?: string;
+}
+
+export const EditableProfileCardHeader = memo(
+    (props: EditableProfileCardHeaderProps) => {
+        const { className } = props;
+        const { t } = useTranslation('profile');
+        const readonly = useSelector(getProfileReadonly);
+        const authData = useSelector(getUserAuthData);
+        const profileData = useSelector(getProfileData);
+        const dispatch = useAppDispatch();
+
+        const canEdit = authData?.id === profileData?.id;
+
+        const onEdit = useCallback(() => {
+            dispatch(profileActions.setReadonly(false));
+        }, [dispatch]);
+
+        const onCancelEdit = useCallback(() => {
+            dispatch(profileActions.cancelEdit());
+        }, [dispatch]);
+
+        const onSave = useCallback(() => {
+            dispatch(updateProfileData());
+        }, [dispatch]);
+
+        return (
+            <HStack
+                max
+                justify="between"
+                className={classNames('', {}, [className])}
+            >
+                <Text title={t('Страница профиля')} />
+                {canEdit && (
+                    <div>
+                        {readonly ? (
+                            <Button
+                                theme={ButtonTheme.BACKGROUND_SECONDARY}
+                                onClick={onEdit}
+                            >
+                                {t('Редактировать')}
+                            </Button>
+                        ) : (
+                            <HStack gap="8">
+                                <Button
+                                    theme={ButtonTheme.BACKGROUND_RED}
+                                    onClick={onCancelEdit}
+                                >
+                                    {t('Отменить')}
+                                </Button>
+                                <Button
+                                    theme={ButtonTheme.BACKGROUND_SECONDARY}
+                                    onClick={onSave}
+                                >
+                                    {t('Сохранить')}
+                                </Button>
+                            </HStack>
+                        )}
+                    </div>
+                )}
+            </HStack>
+        );
+    }
+);
