+++ src/pages/ArticleDetailsPage/ui/ArticleDetailsComments/ArticleDetailsComments.stories.tsx
@@ -1,0 +1,18 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { ArticleDetailsComments } from './ArticleDetailsComments';
+
+export default {
+    title: 'pages/ArticleDetailsPage/ArticleDetailsComments',
+    component: ArticleDetailsComments,
+    argTypes: {
+        backgroundColor: { control: 'color' }
+    }
+} as ComponentMeta<typeof ArticleDetailsComments>;
+
+// eslint-disable-next-line react/jsx-props-no-spreading
+const Template: ComponentStory<typeof ArticleDetailsComments> = (args) => (
+    <ArticleDetailsComments {...args} />
+);
+
+export const Normal = Template.bind({});
+Normal.args = {};
