+++ src/pages/ArticleDetailsPage/ui/ArticleDetailsComments/ArticleDetailsComments.tsx
@@ -1,0 +1,59 @@
+import { useTranslation } from 'react-i18next';
+import { memo, Suspense, useCallback } from 'react';
+import { classNames } from 'shared/lib/helpers/classNames';
+import { TextAlign, Text } from 'shared/ui/Text/Text';
+import { AddCommentForm } from 'features/AddCommentForm';
+import { Loader } from 'shared/ui/Loader/Loader';
+import { CommentList } from 'entities/Comment';
+import { useSelector } from 'react-redux';
+import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
+import { useInitialEffect } from 'shared/lib/hooks/useInitialEffect/useInitialEffect';
+import { VStack } from 'shared/ui/Stack';
+import { getArticleComments } from '../../model/slice/articleDetailsCommentsSlice';
+import { getArticleCommentsIsLoading } from '../../model/selectors/comment/comment';
+import { getArticleRecommendationsIsLoading } from '../../model/selectors/recommendations/recommendations';
+import { addCommentForArticle } from '../../model/services/addCommentForArticle/addCommentForArticle';
+import { fetchCommentsByArticleId } from '../../model/services/fetchCommentsByArticleId/fetchCommentsByArticleId';
+
+interface ArticleDetailsCommentsProps {
+    className?: string;
+    id: string;
+}
+
+export const ArticleDetailsComments = memo(
+    (props: ArticleDetailsCommentsProps) => {
+        const { className, id } = props;
+        const { t } = useTranslation();
+        const dispatch = useAppDispatch();
+
+        const comments = useSelector(getArticleComments.selectAll);
+        const commentsIsLoading = useSelector(getArticleCommentsIsLoading);
+        const recommendationsIsLoading = useSelector(
+            getArticleRecommendationsIsLoading
+        );
+
+        const onSendComment = useCallback(
+            (text: string) => {
+                dispatch(addCommentForArticle(text));
+            },
+            [dispatch]
+        );
+
+        useInitialEffect(() => {
+            dispatch(fetchCommentsByArticleId(id));
+        });
+
+        return (
+            <VStack gap="8" className={classNames('', {}, [className])}>
+                <Text title={t('Комментарии')} align={TextAlign.LEFT} />
+                <Suspense fallback={<Loader />}>
+                    <AddCommentForm onSendComment={onSendComment} />
+                </Suspense>
+                <CommentList
+                    isLoading={commentsIsLoading}
+                    comments={comments}
+                />
+            </VStack>
+        );
+    }
+);
