+++ src/pages/ArticleDetailsPage/ui/ArticleDetailsPage/ArticleDetailsPage.tsx
@@ -1,31 +1,19 @@
-import { ArticleDetails, ArticleList } from 'entities/Article';
-import { CommentList } from 'entities/Comment';
-import { AddCommentForm } from 'features/AddCommentForm';
-import { memo, Suspense, useCallback } from 'react';
+import { ArticleDetails } from 'entities/Article';
+import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import {
     DynamicModuleLoader,
     ReducerList
 } from 'shared/lib/components/DynamicModuleLoader/DynamicModuleLoader';
 import { classNames } from 'shared/lib/helpers/classNames';
-import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
-import { useInitialEffect } from 'shared/lib/hooks/useInitialEffect/useInitialEffect';
-import { Text, TextAlign } from 'shared/ui/Text/Text';
-import { Loader } from 'shared/ui/Loader/Loader';
 import { Page } from 'widgets/Page/Page';
 import { VStack } from 'shared/ui/Stack';
-import { getArticleRecommendations } from '../../model/slice/articleDetailsPageRecommendationsSlice';
-import { fetchCommentsByArticleId } from '../../model/services/fetchCommentsByArticleId/fetchCommentsByArticleId';
-import { addCommentForArticle } from '../../model/services/addCommentForArticle/addCommentForArticle';
-import { getArticleCommentsIsLoading } from '../../model/selectors/comment/comment';
-import { getArticleComments } from '../../model/slice/articleDetailsCommentsSlice';
-import cls from './ArticleDetailsPage.module.scss';
-import { getArticleRecommendationsIsLoading } from '../../model/selectors/recommendations/recommendations';
-import { fetchArticleRecommendations } from '../../model/services/fetchArticleRecommendations/fetchArticleRecommendations';
+import { ArticleRecomendationsList } from 'features/articleRecomendationsList';
 import { articleDetailsPageReducer } from '../../model/slice';
 import { ArticleDetailsPageHeader } from '../ArticleDetailsPageHeader/ArticleDetailsPageHeader';
+import { ArticleDetailsComments } from '../ArticleDetailsComments/ArticleDetailsComments';
+import cls from './ArticleDetailsPage.module.scss';
 
 interface ArticleDetailsPageProps {
     className?: string;
@@ -38,25 +26,6 @@ const reducers: ReducerList = {
 const ArticleDetailsPage = ({ className }: ArticleDetailsPageProps) => {
     const { t } = useTranslation('article-details');
     const { id = '1' } = useParams<{ id: string }>();
-    const dispatch = useAppDispatch();
-    const comments = useSelector(getArticleComments.selectAll);
-    const recommendations = useSelector(getArticleRecommendations.selectAll);
-    const commentsIsLoading = useSelector(getArticleCommentsIsLoading);
-    const recommendationsIsLoading = useSelector(
-        getArticleRecommendationsIsLoading
-    );
-
-    const onSendComment = useCallback(
-        (text: string) => {
-            dispatch(addCommentForArticle(text));
-        },
-        [dispatch]
-    );
-
-    useInitialEffect(() => {
-        dispatch(fetchCommentsByArticleId(id));
-        dispatch(fetchArticleRecommendations());
-    });
 
     if (!id) {
         return (
@@ -72,26 +41,10 @@ const ArticleDetailsPage = ({ className }: ArticleDetailsPageProps) => {
                 <VStack gap="16" max align="normal">
                     <ArticleDetailsPageHeader />
                     <ArticleDetails id={id} />
-                    <Text
-                        title={t('Рекомендуем')}
-                        align={TextAlign.LEFT}
-                        className={cls.recommendationsTitle}
-                    />
-                    <ArticleList
-                        articles={recommendations}
-                        isLoading={recommendationsIsLoading}
+                    <ArticleRecomendationsList
                         className={cls.recommendations}
-                        target="_blank"
-                        skeletonsCount={4}
                     />
-                    <Text title={t('Комментарии')} align={TextAlign.LEFT} />
-                    <Suspense fallback={<Loader />}>
-                        <AddCommentForm onSendComment={onSendComment} />
-                    </Suspense>
-                    <CommentList
-                        isLoading={commentsIsLoading}
-                        comments={comments}
-                    />
+                    <ArticleDetailsComments id={id} />
                 </VStack>
             </Page>
         </DynamicModuleLoader>
