+++ src/pages/ArticlesPage/ui/ArticleInfiniteList/ArticleInfiniteList.tsx
@@ -1,0 +1,38 @@
+import { useTranslation } from 'react-i18next';
+import { memo } from 'react';
+import { useSelector } from 'react-redux';
+import { ArticleList } from 'entities/Article';
+import { Text } from 'shared/ui/Text/Text';
+import { getArticles } from '../../model/slices/articlesPageSlice';
+import {
+    getArticlesPageError,
+    getArticlesPageIsLoading,
+    getArticlesPageView
+} from '../../model/selectors/articlesPageSelectors';
+
+interface ArticleInfiniteListProps {
+    className?: string;
+}
+
+export const ArticleInfiniteList = memo((props: ArticleInfiniteListProps) => {
+    const { className } = props;
+    const { t } = useTranslation();
+
+    const articles = useSelector(getArticles.selectAll);
+    const isLoading = useSelector(getArticlesPageIsLoading);
+    const view = useSelector(getArticlesPageView);
+    const error = useSelector(getArticlesPageError);
+
+    if (error) {
+        return <Text title={t('Произошла ошибка при загрузке статей')} />;
+    }
+
+    return (
+        <ArticleList
+            view={view}
+            articles={articles}
+            isLoading={isLoading}
+            className={className}
+        />
+    );
+});
