--- src/pages/ProfilePage/ui/ProfilePageHeader/ProfilePageHeader.module.scss
@@ -1,3 +1,0 @@
-.ProfilePageHeader {
-    padding: 0 20px;
-}
