--- src/pages/ProfilePage/ui/ProfilePageHeader/ProfilePageHeader.tsx
@@ -1,79 +1,0 @@
-import { useTranslation } from 'react-i18next';
-import { Button, ButtonTheme } from 'shared/ui/Button/Button';
-import { classNames } from 'shared/lib/helpers/classNames';
-import { Text } from 'shared/ui/Text/Text';
-import { useSelector } from 'react-redux';
-import {
-    getProfileData,
-    getProfileReadonly,
-    profileActions,
-    updateProfileData
-} from 'entities/Profile';
-import { useCallback } from 'react';
-import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
-import { getUserAuthData } from 'entities/User';
-import { HStack } from 'shared/ui/Stack';
-import cls from './ProfilePageHeader.module.scss';
-
-interface ProfilePageHeaderProps {
-    className?: string;
-}
-
-export const ProfilePageHeader = ({ className }: ProfilePageHeaderProps) => {
-    const { t } = useTranslation('profile');
-    const readonly = useSelector(getProfileReadonly);
-    const authData = useSelector(getUserAuthData);
-    const profileData = useSelector(getProfileData);
-    const dispatch = useAppDispatch();
-
-    const canEdit = authData?.id === profileData?.id;
-
-    const onEdit = useCallback(() => {
-        dispatch(profileActions.setReadonly(false));
-    }, [dispatch]);
-
-    const onCancelEdit = useCallback(() => {
-        dispatch(profileActions.cancelEdit());
-    }, [dispatch]);
-
-    const onSave = useCallback(() => {
-        dispatch(updateProfileData());
-    }, [dispatch]);
-
-    return (
-        <HStack
-            max
-            justify="between"
-            className={classNames(cls.ProfilePageHeader, {}, [className])}
-        >
-            <Text title={t('Страница профиля')} />
-            {canEdit && (
-                <div>
-                    {readonly ? (
-                        <Button
-                            theme={ButtonTheme.BACKGROUND_SECONDARY}
-                            onClick={onEdit}
-                        >
-                            {t('Редактировать')}
-                        </Button>
-                    ) : (
-                        <HStack gap="8">
-                            <Button
-                                theme={ButtonTheme.BACKGROUND_RED}
-                                onClick={onCancelEdit}
-                            >
-                                {t('Отменить')}
-                            </Button>
-                            <Button
-                                theme={ButtonTheme.BACKGROUND_SECONDARY}
-                                onClick={onSave}
-                            >
-                                {t('Сохранить')}
-                            </Button>
-                        </HStack>
-                    )}
-                </div>
-            )}
-        </HStack>
-    );
-};
